+++ systemimage/tests/test_winner.py
@@ -29,20 +29,12 @@ from systemimage.config import config
 from systemimage.download import get_files
 from systemimage.gpg import SignatureError
 from systemimage.helpers import temporary_directory
-from systemimage.logging import initialize
 from systemimage.scores import WeightedScorer
 from systemimage.state import ChecksumError, State
 from systemimage.testing.helpers import (
     copy, get_index, make_http_server, makedirs, setup_index,
     setup_keyring_txz, setup_keyrings, sign, test_data_path,
     testable_configuration)
-
-
-def setUpModule():
-    # BAW 2013-06-17: For correctness, this really should be put in all
-    # test_*.py modules, or in a global test runner.  As it is, this only
-    # quiets the logging output for tests in this module and later.
-    initialize(verbosity=3)
 
 
 class TestWinnerDownloads(unittest.TestCase):